dataloaders cleanup
