--- engine/dataloaders.py
@@ -1,630 +1,0 @@
-import os
-import numpy as np
-import torch
-import random
-import datasets
-import torch.distributed as dist
-import torch.nn.functional as F
-import re
-
-from torch.utils.data import DataLoader
-from torch.utils.data.distributed import DistributedSampler
-from torch.nn.utils.rnn import pad_sequence
-from config import TrainingStage
-from logger import logger
-
-
-def load_tokens(filename):
-    npt = np.load(filename, allow_pickle=False)
-    if npt.dtype != np.int64:
-        npt = npt.astype(np.int64)
-    ptt = torch.from_numpy(npt)
-    return ptt
-
-def pad_batch_to_multiple_of(*, sequences, padding_value, multiple, max_length=None):
-    padded = pad_sequence(
-        sequences,
-        batch_first=True,
-        padding_value=padding_value
-    )
-
-    if multiple is None or multiple <= 1:
-        return padded
-
-    seq_len = padded.size(1)
-    target_len = ((seq_len + multiple - 1) // multiple) * multiple
-    if max_length is not None:
-        target_len = min(target_len, max_length)
-
-    remaining = target_len - seq_len
-    if remaining <= 0:
-        return padded
-    return F.pad(padded, (0, remaining), value=padding_value)
-
-class PretrainingDataLoader:
-    def __init__(
-        self,
-        batch_size,
-        sequence_length,
-        is_master_process,
-        ddp_rank,
-        ddp_world_size,
-        data_root,
-        split,
-        use_shuffle=False
-    ):
-        self.B = batch_size
-        self.S = sequence_length
-        self.is_master_process = is_master_process
-        self.ddp_rank = ddp_rank
-        self.ddp_world_size = ddp_world_size
-        self.data_root = data_root
-        assert split in {'train', 'val'}
-        self.split = split
-        self.use_shuffle = use_shuffle
-        self.total_tokens = None
-
-        split_root = os.path.join(data_root, split)
-        assert os.path.isdir(split_root), f'missing split dir: {split_root}'
-
-        target_pattern = re.compile(r'^data_(\d+)\.npy$')
-        valid_shards = []
-        for file_name in os.listdir(split_root):
-            match = target_pattern.match(file_name)
-            if match:
-                valid_shards.append((int(match.group(1)), os.path.join(split_root, file_name)))
-
-        valid_shards.sort(key=lambda x: x[0])
-        indexes = [i for i, _ in valid_shards]
-        assert indexes == list(range(len(valid_shards))), f'Shard sequence is broken: {indexes}'
-
-        self.shards = [shard_path for _, shard_path in valid_shards]
-        assert self.shards, f'no shards found in split {split}'
-
-        logger.info(f'found {len(self.shards)} shards for split {split}')
-
-        self.validate_shards_size()
-        self.reset()
-
-    def validate_shards_size(self):
-        required_tokens = self.B * self.S * self.ddp_world_size + 1
-        for shard_path in self.shards:
-            shard = np.load(shard_path, mmap_mode='r', allow_pickle=False)
-            shard_len = int(shard.shape[0])
-
-            if shard_len < required_tokens:
-                raise ValueError(
-                    f'Shard is too small for distributed training: {shard_path}. '
-                    f'Need a minimum of {required_tokens} tokens, got {shard_len}. '
-                    f'B={self.B}, S={self.S}, world_size={self.ddp_world_size}.'
-                )
-
-    def calculate_max_tokens(self):
-        if self.total_tokens:
-            return self.total_tokens
-
-        def _calculate():
-            total = 0
-            for path in self.shards:
-                shard = np.load(path, mmap_mode='r', allow_pickle=False)
-                total += int(shard.shape[0])
-                del shard
-            return total
-
-        if self.ddp_world_size <= 1 or not dist.is_available() or not dist.is_initialized():
-            return _calculate()
-
-        total_tokens = None
-        object_list_to_sync = [total_tokens]
-        if self.is_master_process:
-            object_list_to_sync[0] = _calculate()
-        dist.broadcast_object_list(object_list_to_sync, src=0)
-        total_tokens = int(object_list_to_sync[0])
-        self.total_tokens = total_tokens
-        return total_tokens
-
-    def sync_shuffle_shards(self):
-        if not self.use_shuffle:
-            return
-
-        if self.ddp_world_size <= 1 or not dist.is_available() or not dist.is_initialized():
-            random.shuffle(self.shards)
-            return
-
-        # create the indexes and shuffle
-        target_indexes = list(range(len(self.shards)))
-        if self.is_master_process:
-            random.shuffle(target_indexes)
-
-        # synchronize the shuffle
-        object_list_to_sync = [target_indexes]
-        dist.broadcast_object_list(object_list_to_sync, src=0)
-        order = object_list_to_sync[0]
-        self.shards = [self.shards[i] for i in order]
-
-    def reset(self):
-        self.current_shard = 0
-        self.sync_shuffle_shards()
-        self.tokens = load_tokens(self.shards[self.current_shard])
-        if torch.cuda.is_available():
-            self.tokens = self.tokens.pin_memory()
-        self.current_position = self.B * self.S * self.ddp_rank
-
-    def state_dict(self):
-        return {
-            'shards' : list(self.shards),
-            'current_shard' : self.current_shard,
-            'current_position' : self.current_position
-        }
-
-    def load_state_dict(self, state):
-        self.shards = state['shards']
-        self.current_shard = state['current_shard']
-        self.tokens = load_tokens(self.shards[self.current_shard])
-        if torch.cuda.is_available():
-            self.tokens = self.tokens.pin_memory()
-        self.current_position = state['current_position']
-
-    def next_batch(self):
-        B, S = self.B, self.S
-        buf = self.tokens[self.current_position : self.current_position+B*S+1]
-        x = (buf[:-1]).view(B, S)
-        y = (buf[1:]).view(B, S)
-        self.current_position += B * S * self.ddp_world_size
-        if self.current_position + (B * S * self.ddp_world_size + 1) > len(self.tokens):
-            self.current_shard = (self.current_shard + 1) % len(self.shards)
-            if self.current_shard == 0:
-                self.sync_shuffle_shards()
-            self.tokens = load_tokens(self.shards[self.current_shard])
-            if torch.cuda.is_available():
-                self.tokens = self.tokens.pin_memory()
-            self.current_position = self.B * self.S * self.ddp_rank
-        return x, y
-
-class InstructDataLoader:
-    def __init__(
-        self,
-        batch_size,
-        sequence_length,
-        is_master_process,
-        ddp_rank,
-        ddp_world_size,
-        data_root,
-        split,
-        use_shuffle,
-        pad_id,
-        drop_last,
-        number_of_cpu_processes,
-        ignore_index
-    ):
-        self.B = batch_size
-        self.S = sequence_length
-        self.is_master_process = is_master_process
-        self.ddp_rank = ddp_rank
-        self.ddp_world_size = ddp_world_size
-        self.total_tokens = None
-        self.number_of_cpu_processes = number_of_cpu_processes
-        self.ignore_index = ignore_index
-
-        dataset = datasets.load_from_disk(os.path.join(data_root, split))
-        logger.info(f'found {len(dataset)} examples for {split}')
-
-        self.is_master_process = is_master_process
-
-        assert isinstance(dataset, datasets.Dataset)
-
-        self.sampler = DistributedSampler(
-            dataset,
-            num_replicas=ddp_world_size,
-            rank=ddp_rank,
-            shuffle=use_shuffle,
-            drop_last=drop_last
-        )
-
-        def collate(examples):
-            ids = []
-            labels = []
-
-            for i, e in enumerate(examples):
-                input_ids = torch.tensor(e['input_ids'], dtype=torch.long)
-                target_labels = torch.tensor(e['labels'], dtype=torch.long)
-
-                if input_ids.numel() == 0:
-                    raise ValueError(f'Empty input_ids in collate example {i}: {e}')
-                if target_labels.numel() == 0:
-                    raise ValueError(f'Empty labels in collate example {i}: {e}')
-                if input_ids.numel() != target_labels.numel():
-                    raise ValueError(
-                        f'input_ids/labels length do not match in collate example {i}: '
-                        f'{input_ids.numel()} vs {target_labels.numel()}'
-                    )
-                if (target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No labels before truncation in collate example {i}: {e}')
-                if input_ids.numel() > sequence_length:
-                    input_ids = input_ids[-sequence_length:]
-                    target_labels = target_labels[-sequence_length:]
-                if (input_ids != int(pad_id)).sum().item() == 0:
-                    raise ValueError(f'Input ids after truncation are all pad tokens in collate example {i}: {e}')
-                if (target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No labels after truncation in collate example {i}: {e}')
-
-                ids.append(input_ids)
-                labels.append(target_labels)
-
-            ids = pad_batch_to_multiple_of(
-                sequences=ids,
-                padding_value=int(pad_id),
-                multiple=8,
-                max_length=sequence_length,
-            )
-            labels = pad_batch_to_multiple_of(
-                sequences=labels,
-                padding_value=self.ignore_index,
-                multiple=8,
-                max_length=sequence_length,
-            )
-
-            return ids, labels
-
-        self._dataloader = DataLoader(
-            dataset,
-            batch_size=batch_size,
-            sampler=self.sampler,
-            collate_fn=collate,
-            drop_last=drop_last,
-            pin_memory=True
-        )
-        self._iterator = iter(self._dataloader)
-
-    def next_batch(self):
-        try:
-            return next(self._iterator)
-        except StopIteration:
-            self.reset()
-            return next(self._iterator)
-
-    def reset(self):
-        if hasattr(self.sampler, 'set_epoch'):
-            update_epoch = 0
-            if hasattr(self.sampler, 'epoch'):
-                update_epoch = self.sampler.epoch + 1
-            self.sampler.set_epoch(update_epoch)
-        self._iterator = iter(self._dataloader)
-
-    def num_examples(self):
-        return len(self._dataloader.dataset)
-
-    def calculate_max_tokens(self):
-        if self.total_tokens:
-            return self.total_tokens
-
-        def _calculate():
-            return sum(self._dataloader.dataset.map(
-                lambda ex: {'len': len(ex['input_ids'])},
-                num_proc=self.number_of_cpu_processes,
-                remove_columns=[],
-                desc='Calculating number of tokens'
-            )['len'])
-
-        if self.ddp_world_size <= 1 or not dist.is_available() or not dist.is_initialized():
-            return _calculate()
-
-        total_tokens = None
-        object_list_to_sync = [total_tokens]
-        if self.is_master_process:
-            object_list_to_sync[0] = _calculate()
-        dist.broadcast_object_list(object_list_to_sync, src=0)
-        total_tokens = int(object_list_to_sync[0])
-        self.total_tokens = total_tokens
-        return total_tokens
-
-    def state_dict(self):
-        return {'epoch': getattr(self.sampler, 'epoch', 0)}
-
-    def load_state_dict(self, state):
-        if 'epoch' not in state:
-            logger.warn('"epoch" not present, starting fresh dataloader (most likely transition from pretraining to SFT).')
-            return
-        epoch = state['epoch']
-        self.sampler.set_epoch(epoch)
-        self.sampler.epoch = epoch
-        self._iterator = iter(self._dataloader)
-
-class DirectPreferenceOptimizationDataLoader:
-    def __init__(
-        self,
-        batch_size,
-        sequence_length,
-        is_master_process,
-        ddp_rank,
-        ddp_world_size,
-        data_root,
-        split,
-        use_shuffle,
-        pad_id,
-        drop_last,
-        number_of_cpu_processes,
-        ignore_index
-    ):
-        self.B = batch_size
-        self.S = sequence_length
-        self.is_master_process = is_master_process
-        self.ddp_rank = ddp_rank
-        self.ddp_world_size = ddp_world_size
-        self.total_tokens = None
-        self.number_of_cpu_processes = number_of_cpu_processes
-        self.ignore_index = ignore_index
-
-        dataset = datasets.load_from_disk(os.path.join(data_root, split))
-        logger.info(f'found {len(dataset)} examples for {split}')
-
-        self.is_master_process = is_master_process
-
-        assert isinstance(dataset, datasets.Dataset)
-
-        self.sampler = DistributedSampler(
-            dataset,
-            num_replicas=ddp_world_size,
-            rank=ddp_rank,
-            shuffle=use_shuffle,
-            drop_last=drop_last
-        )
-
-        def collate(examples):
-            chosen_ids = []
-            chosen_labels = []
-            rejected_ids = []
-            rejected_labels = []
-
-            for i, e in enumerate(examples):
-                chosen_input_ids = torch.tensor(e['chosen_input_ids'], dtype=torch.long)
-                chosen_target_labels = torch.tensor(e['chosen_labels'], dtype=torch.long)
-                rejected_input_ids = torch.tensor(e['rejected_input_ids'], dtype=torch.long)
-                rejected_target_labels = torch.tensor(e['rejected_labels'], dtype=torch.long)
-
-                if chosen_input_ids.numel() == 0:
-                    raise ValueError(f'Empty chosen_input_ids in collate example {i}: {e}')
-                if chosen_target_labels.numel() == 0:
-                    raise ValueError(f'Empty chosen_labels in collate example {i}: {e}')
-                if rejected_input_ids.numel() == 0:
-                    raise ValueError(f'Empty rejected_input_ids in collate example {i}: {e}')
-                if rejected_target_labels.numel() == 0:
-                    raise ValueError(f'Empty rejected_labels in collate example {i}: {e}')
-                if chosen_input_ids.numel() != chosen_target_labels.numel():
-                    raise ValueError(
-                        f'chosen input_ids/labels length do not match in collate example {i}: '
-                        f'{chosen_input_ids.numel()} vs {chosen_target_labels.numel()}'
-                    )
-                if rejected_input_ids.numel() != rejected_target_labels.numel():
-                    raise ValueError(
-                        f'rejected input_ids/labels length do not match in collate example {i}: '
-                        f'{rejected_input_ids.numel()} vs {rejected_target_labels.numel()}'
-                    )
-                if (chosen_target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No chosen labels before truncation in collate example {i}: {e}')
-                if (rejected_target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No rejected labels before truncation in collate example {i}: {e}')
-
-                if chosen_input_ids.numel() > sequence_length:
-                    chosen_input_ids = chosen_input_ids[-sequence_length:]
-                    chosen_target_labels = chosen_target_labels[-sequence_length:]
-                if rejected_input_ids.numel() > sequence_length:
-                    rejected_input_ids = rejected_input_ids[-sequence_length:]
-                    rejected_target_labels = rejected_target_labels[-sequence_length:]
-
-                if (chosen_input_ids != int(pad_id)).sum().item() == 0:
-                    raise ValueError(f'Chosen input ids after truncation are all pad tokens in collate example {i}: {e}')
-                if (rejected_input_ids != int(pad_id)).sum().item() == 0:
-                    raise ValueError(f'Rejected input ids after truncation are all pad tokens in collate example {i}: {e}')
-                if (chosen_target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No chosen labels after truncation in collate example {i}: {e}')
-                if (rejected_target_labels != self.ignore_index).sum().item() == 0:
-                    raise ValueError(f'No rejected labels after truncation in collate example {i}: {e}')
-
-                chosen_ids.append(chosen_input_ids)
-                chosen_labels.append(chosen_target_labels)
-                rejected_ids.append(rejected_input_ids)
-                rejected_labels.append(rejected_target_labels)
-
-            chosen_ids = pad_batch_to_multiple_of(
-                sequences=chosen_ids,
-                padding_value=int(pad_id),
-                multiple=8,
-                max_length=sequence_length,
-            )
-            chosen_labels = pad_batch_to_multiple_of(
-                sequences=chosen_labels,
-                padding_value=self.ignore_index,
-                multiple=8,
-                max_length=sequence_length,
-            )
-            rejected_ids = pad_batch_to_multiple_of(
-                sequences=rejected_ids,
-                padding_value=int(pad_id),
-                multiple=8,
-                max_length=sequence_length,
-            )
-            rejected_labels = pad_batch_to_multiple_of(
-                sequences=rejected_labels,
-                padding_value=self.ignore_index,
-                multiple=8,
-                max_length=sequence_length,
-            )
-
-            return chosen_ids, chosen_labels, rejected_ids, rejected_labels
-
-        self._dataloader = DataLoader(
-            dataset,
-            batch_size=batch_size,
-            sampler=self.sampler,
-            collate_fn=collate,
-            drop_last=drop_last,
-            pin_memory=True
-        )
-        self._iterator = iter(self._dataloader)
-
-    def next_batch(self):
-        try:
-            return next(self._iterator)
-        except StopIteration:
-            self.reset()
-            return next(self._iterator)
-
-    def reset(self):
-        if hasattr(self.sampler, 'set_epoch'):
-            update_epoch = 0
-            if hasattr(self.sampler, 'epoch'):
-                update_epoch = self.sampler.epoch + 1
-            self.sampler.set_epoch(update_epoch)
-        self._iterator = iter(self._dataloader)
-
-    def num_examples(self):
-        return len(self._dataloader.dataset)
-
-    def calculate_max_tokens(self):
-        if self.total_tokens:
-            return self.total_tokens
-
-        def _calculate():
-            return sum(self._dataloader.dataset.map(
-                lambda ex: {'len': len(ex['chosen_input_ids']) + len(ex['rejected_input_ids'])},
-                num_proc=self.number_of_cpu_processes,
-                remove_columns=[],
-                desc='Calculating number of tokens'
-            )['len'])
-
-        if self.ddp_world_size <= 1 or not dist.is_available() or not dist.is_initialized():
-            return _calculate()
-
-        total_tokens = None
-        object_list_to_sync = [total_tokens]
-        if self.is_master_process:
-            object_list_to_sync[0] = _calculate()
-        dist.broadcast_object_list(object_list_to_sync, src=0)
-        total_tokens = int(object_list_to_sync[0])
-        self.total_tokens = total_tokens
-        return total_tokens
-
-    def state_dict(self):
-        return {'epoch': getattr(self.sampler, 'epoch', 0)}
-
-    def load_state_dict(self, state):
-        if 'epoch' not in state:
-            if self.is_master_process:
-                logger.warn('"epoch" not present, starting fresh dataloader (most likely transition from pretraining to DPO).')
-            return
-        epoch = state['epoch']
-        self.sampler.set_epoch(epoch)
-        self.sampler.epoch = epoch
-        self._iterator = iter(self._dataloader)
-
-def init_data_loaders(
-    batch_size,
-    sequence_length,
-    is_master_process,
-    ddp_rank,
-    ddp_world_size,
-    data_root,
-    training_stage,
-    number_of_cpu_processes,
-    ignore_index,
-    pad_id=None,
-    validation_only=False
-):
-    train_loader = None
-    if training_stage == TrainingStage.PRETRAINING:
-        logger.section('Pretraining Data Loaders')
-
-        if validation_only is False:
-            train_loader = PretrainingDataLoader(
-                batch_size=batch_size,
-                sequence_length=sequence_length,
-                is_master_process=is_master_process,
-                ddp_rank=ddp_rank,
-                ddp_world_size=ddp_world_size,
-                data_root=data_root,
-                split='train',
-                use_shuffle=True
-            )
-        val_loader = PretrainingDataLoader(
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            is_master_process=is_master_process,
-            ddp_rank=ddp_rank,
-            ddp_world_size=ddp_world_size,
-            data_root=data_root,
-            split='val',
-            use_shuffle=False
-        )
-    elif training_stage == TrainingStage.INSTRUCT:
-        assert pad_id is not None
-
-        logger.section('Instruct Finetuning Data Loaders')
-
-        if validation_only is False:
-            train_loader = InstructDataLoader(
-                batch_size=batch_size,
-                sequence_length=sequence_length,
-                is_master_process=is_master_process,
-                ddp_rank=ddp_rank,
-                ddp_world_size=ddp_world_size,
-                data_root=data_root,
-                split='train',
-                use_shuffle=True,
-                pad_id=pad_id,
-                drop_last=True,
-                number_of_cpu_processes=number_of_cpu_processes,
-                ignore_index=ignore_index
-            )
-        val_loader = InstructDataLoader(
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            is_master_process=is_master_process,
-            ddp_rank=ddp_rank,
-            ddp_world_size=ddp_world_size,
-            data_root=data_root,
-            split='val',
-            use_shuffle=False,
-            pad_id=pad_id,
-            drop_last=False,
-            number_of_cpu_processes=number_of_cpu_processes,
-            ignore_index=ignore_index
-        )
-    elif training_stage == TrainingStage.DPO:
-        assert pad_id is not None
-
-        logger.section('Direct Preference Optimization Data Loaders')
-
-        if validation_only is False:
-            train_loader = DirectPreferenceOptimizationDataLoader(
-                batch_size=batch_size,
-                sequence_length=sequence_length,
-                is_master_process=is_master_process,
-                ddp_rank=ddp_rank,
-                ddp_world_size=ddp_world_size,
-                data_root=data_root,
-                split='train',
-                use_shuffle=True,
-                pad_id=pad_id,
-                drop_last=True,
-                number_of_cpu_processes=number_of_cpu_processes,
-                ignore_index=ignore_index
-            )
-        val_loader = DirectPreferenceOptimizationDataLoader(
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            is_master_process=is_master_process,
-            ddp_rank=ddp_rank,
-            ddp_world_size=ddp_world_size,
-            data_root=data_root,
-            split='val',
-            use_shuffle=False,
-            pad_id=pad_id,
-            drop_last=False,
-            number_of_cpu_processes=number_of_cpu_processes,
-            ignore_index=ignore_index
-        )
-    else:
-        raise ValueError('Invalid training stage for dataloader')
-
-    return train_loader, val_loader
